set jwt authentication

+++ controllers/api/v1/posts_api.js
@@ -23,20 +23,21 @@ module.exports.destroy = async (req, res) => {
     let post = await Post.findById(req.params.id);
 
     //.id means converting the object id into string
-    //   if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    return res.json(200, {
-      message: "Posts and associated comments deleted successfully!",
-    });
-
-    //   } else {
-    //     req.flash("error", "You cannot delete this post");
-    //     return res.redirect("back");
-    //   }
+      return res.json(200, {
+        message: "Posts and associated comments deleted successfully!",
+      });
+    } else {
+      return res.json(401, {
+        message: "You cannot delete this post!",
+      });
+    }
   } catch (err) {
+    console.log("*******", err);
     return res.json(500, {
       message: "Internal Server Error",
     });
